Integração com Flask-Login (WIP) - Integração do módulo de login com a biblioteca Flask-Login (ainda requer adaptações adicionais, como a modularização) - Essa adaptação permitirá funcionalidades como RememberMe, sessão de usuário e redirecionamento automático de páginas protegidas por login
