+++ requirements.txt
@@ -2,6 +2,7 @@ Flask
 flask_sqlalchemy
 flask_migrate
 flask_wtf
+flask_login
 
 wtforms
 email_validator
