+++ skoolit/usuarios/templates/usuarios/listar.html
@@ -12,7 +12,7 @@
 	{% for usuario in usuarios %}
 		<li>{{usuario.nome}}, email: {{usuario.email}} ({{usuario.papel}}) | 
 			<a href="{{url_for('usuarios.atualizar', id=usuario.id)}}">Editar</a> |
-			{% if g.user.id != usuario.id %}
+			{% if current_user.id != usuario.id %}
 			<a href="{{url_for('usuarios.excluir', id=usuario.id)}}">Excluir</a> |</li>
 			{% endif %}
